Add files via upload
Assignment: PDF Data Extraction and Rapid Prototyping

Objective:

The goal of this assignment is to assess your ability to extract data from a PDF and to rapidly prototype an application based on given requirements.

Part 1: Data Extraction

Task:

Given a PDF document of an invoice, your task is to create a Python script that can extract key-value pairs from the document and save this data into a CSV file. The PDF will contain both header data and tabular data. Your script should be able to handle both types of data.

Task:Part 2: Rapid Prototyping

Design and develop a simple web application that allows users to upload a PDF file, runs your script from Part 1 to extract data, and displays the extracted data on the webpage.